% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{events}
\alias{events}
\title{Event data for hand-picked events that affected
traffic flow in the region of Tyne and Wear in the year 2018.}
\format{A list of lists, each containing the following event data:
\describe{
\item{\code{nodes}}{location ids potentially affected by the incident}
\item{\code{flows_l}}{Tibble containing the number of vehicles passing
through each of the locations during each time period}
\item{\code{flows_od}}{Tibble containing the number, speed statistics
and other information of vehicles passing between each pair of locations
during each time period}
\item{\code{spatial}}{List of spatial data, simple feature dataframes,
cropped to the area of interest monitored locations, shortest paths
between locations, the enclosing arterial and primary road networks
 and nearby amenities}
\item{\code{network}}{asymptotic flow network computed over the chosen
spatio-temporal window}
}}
\usage{
events
}
\description{
Includes traffic flow data acquired from a network of Automatic
Number Plate Recognition (ANPR) cameras and computed via the
anprx python package. Using the anprflows R package, a spatio-temporal
window is used to capture relevant traffic and spatial data
around the epicentre of the traffic event.
}
\keyword{datasets}
